backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
pydantic-settings>=2.2,<3.0
python-multipart>=0.0.9,<1.0
openpyxl>=3.1,<4.0
